fixing sensor macro

+++ software/react_compiler/src/main/scala/react/runtime/Scheduler.scala
@@ -4,7 +4,7 @@ import scala.collection.mutable.PriorityQueue
 import scala.math.Ordering.Implicits
 
 class ScheduledTask(var expires: Long,
-                    val period: Long,
+                    val period: Int,
                     val fct: () => Unit
                    ) extends java.lang.Comparable[ScheduledTask] {
   def compareTo(other: ScheduledTask) = other.expires.compareTo(expires)
@@ -17,7 +17,7 @@ class Scheduler {
 
   private def now = java.lang.System.currentTimeMillis()
 
-  def addTask(period: Long, fct: () => Unit) {
+  def addTask(period: Int, fct: () => Unit) {
     assert(period >= 1, "period needs to be at least 1ms")
     val task = new ScheduledTask(now + period, period, fct)
     queue.enqueue(task)
